+++ drive_backup.py
@@ -1,0 +1,45 @@
+
+import os
+import zipfile
+from google.oauth2.credentials import Credentials
+from google_auth_oauthlib.flow import InstalledAppFlow
+from googleapiclient.discovery import build
+from googleapiclient.errors import HttpError
+from googleapiclient.http import MediaFileUpload
+from google.auth.transport.requests import Request
+import logging
+
+from absl import flags
+
+FLAGS = flags.FLAGS
+
+flags.DEFINE_string("backup_directory", ".", "The directory to backup.")
+flags.DEFINE_string("zip_filename", "backup.zip", "The filename for the zipped backup.")
+
+def zipdir(path, ziph):
+    # Zip the directory
+    for root, dirs, files in os.walk(path):
+        for file in files:
+            ziph.write(os.path.join(root, file), os.path.relpath(os.path.join(root, file), path))
+
+#User authentication using google API
+def authenticate():
+    creds = None
+    token_path = "token.json"
+    credentials_path = "client_secret.json"
+
+    if os.path.exists(token_path):
+        creds = Credentials.from_authorized_user_file(token_path)
+
+    if not creds or not creds.valid:
+        if creds and creds.expired and creds.refresh_token:
+            creds.refresh(Request())
+        else:
+            flow = InstalledAppFlow.from_client_secrets_file(credentials_path,
+                                                             ["https://www.googleapis.com/auth/drive.file"])
+            creds = flow.run_local_server(port=0)
+
+        with open(token_path, "w") as token:
+            token.write(creds.to_json())
+
+    return creds
